Create README.md
Welcome to my work on problemset1 in cs50 In this problem I will show you the methodology of thinking in these types of problems which are very easy but it depends on some logic. Let's begin with the concept of problem, like many problems you must draw a specific form with stars. Here you must know that it's very difficult to solve this problem without a small hint which is the division of the pyramid (the specific form in this problem) to two pieces. "Stars'' and "fake points", why fake points? Because it's easier to see the free spaces replaced by points in the beginning then delete them and draw our pyramid. How is it done? First we must draw line by line using iteration, a for loop for the first piece of points then another for stars after that we must separate the two semi-pyramids with two points for each line, then we continue iterating through the second semi-pyramid. Here we remove these points and the difficult part is done. But how is the iterating done? Sure the iteration depends on the parameter of the function, how is that? Usually we can't make a global base for iterating, but you must see the number of stars and points how its changing and make a numerical sequence and use it ,thus we solved the big part of problem and the smallest left, which depends on the parameter (no - numbers , the parameter is between 5 and 9).